调整文件结构

+++ ark-identify/ark-identify-infrastructure/src/main/java/com/ark/identify/infrastucture/persistence/unique_user/model/UniqueUserPO.java
@@ -2,10 +2,10 @@ package com.ark.identify.infrastucture.persistence.unique_user.model;
 
 import com.ark.base.object.BaseDO;
 import com.baomidou.mybatisplus.annotation.TableName;
-import java.io.Serializable;
-import java.time.LocalDateTime;
 import lombok.Getter;
 import lombok.Setter;
+
+import java.time.LocalDateTime;
 
 /**
  * <p>
@@ -35,7 +35,7 @@ public class UniqueUserPO extends BaseDO {
     /**
      * 部门id
      */
-    private Integer departmentId;
+    private Long departmentId;
 
     /**
      * 创建人
